+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - **Fix** : Apple Pay failure occurring during timeout responses scenarios.
 - **Fix** : Multibanco SIBS SPG integration -> Order API
 - **Fix** : Handle "on hold" orders during authorization status update (status 116)
+- **Fix** : Improved RestoreBasketObserver to handle missing order during early HiPay notifications.
 
 ## 1.27.6